Boil down the implementation to pure access.
