+++ src/xcode/ENA/HealthCertificateToolkit/Sources/HealthCertificateToolkit/Validation/ValidationError.swift
@@ -4,7 +4,6 @@
 
 import Foundation
 
-enum RuleValidationError: Error {
-    case ONBOARDED_COUNTRIES_SERVER_ERROR
-    case HC_CBOR_DECODING_FAILED(Error?)
+public enum RuleValidationError: Error {
+    case CBOR_DECODING_FAILED(Error?)
 }
